+++ src/v3/label/LabelReducer.java
@@ -2,7 +2,7 @@ package v3.label;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
@@ -10,9 +10,9 @@ import org.apache.hadoop.mapreduce.Reducer;
  * do nothing
  * 
  */
-public class LabelReducer extends Reducer<IntWritable, Text, IntWritable, Text> {
+public class LabelReducer extends Reducer<LongWritable, Text, LongWritable, Text> {
 
-	public void reduce(IntWritable key, Iterable<Text> values, Context context)
+	public void reduce(LongWritable key, Iterable<Text> values, Context context)
 			throws IOException, InterruptedException {
 		for (Text value : values)
 			context.write(key, value);
